backend/requirements.txt
# Core web framework
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
sse-starlette==1.6.5
starlette==0.27.0
anyio==3.7.1
httpx==0.24.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
alembic==1.12.1
neo4j==5.14.0

# RAG components
sentence-transformers==2.2.2
tiktoken==0.5.2
faiss-cpu==1.7.4

# LLM providers
openai==1.10.0
# use OpenAI client for DeepSeek API now

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
